labs/week1/practice/nguyen/full-stack-gpt/requirements.txt
langchain-openai>=1.0.0
langchain-core>=1.0.0
langchain-community>=0.4.0
langchain-text-splitters>=1.0.0
langchain-classic>=1.0.0
tiktoken>=0.8.0
faiss-cpu
streamlit
python-dotenv>=1.0.0
jupyter
ipykernel
